remove synchronous file io
`getMIMEFromPath` previously did a `ReadChunk.sync`, which throws an
exception if you're out of file descriptors (or some other error
occurs). Changing this to asynchronous with an empty mime type on
failure. It's now pretty easy to handle EMFILE errors with something
like graceful-fs (which catches and delays async io EMFILE errors).