Use Buffer.isBuffer() instead of constructor check
Ran in to an issue where using jimp/browser was failing in an odd way and returning false for the buffer checks even though the actual value was a buffer. I don't know if it was the particular environment (babel + webpack) but changing it to use Buffer.isBuffer() instead of comparing the constructor function seemed to do the trick for me.