Write using saved mime type (#403)
* Write using saved mime type

This allows filenames without a suffix to save properly using the already established MIME type. Closes #114. Though, this doesn't add any functionality for changing the mime type as referenced in the issue #114.

Note: MIME.lookup(path) is actually already being used to obtain the value that this.getMIME() references.

* MIME defaults to original MIME if not set

If neither is set getMIME defaults to type PNG

* Revert

* Inferred MIME types

Added an example to the code as well as mentioning that it all defaults to PNG

* correct function

* add test