Add more details to the Align Modes (#670)
I was having trouble recently figuring out how to center some text on an image and I kept thinking that the `Jimp.HORIZONTAL_ALIGN_CENTER` meant that the cursor position of the text object would be at its horizontal center, not that the position of the x-axis on the image to the center. After a bunch of trial and error, I finally realized that these align modes change the position of the axis that you're specifying.

I figured that adding this table and adjusting the example for `print` might help other users who may run into this issue.